pyproject.toml
[tool.pdm]

[project]
name = "lco"
version = "0.1.0"
description = "A light and global config for research, MVPs and prototype projects."
authors = [
    {name = "Christoph Minixhofer", email = "christoph.minixhofer@gmail.com"},
]
dependencies = [
    "pyyaml>=6.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
